+++ addie/mantid_handler/mantid_thread.py
@@ -5,7 +5,7 @@ import os
 from mantid.simpleapi import *
 import mantid
 
-from addie.processing_idl.mantid_script_handler import MantidScriptHandler
+from addie.processing.idl.mantid_script_handler import MantidScriptHandler
 
 
 class MantidThread(QThread):
