+++ addie/processing/idl/create_ndsum_file.py
@@ -1,7 +1,7 @@
 from __future__ import (absolute_import, division, print_function)
 import os
-from addie.processing_idl.table_handler import TableHandler
-from addie.processing_idl.step2_gui_handler import Step2GuiHandler
+from addie.processing.idl.table_handler import TableHandler
+from addie.processing.idl.step2_gui_handler import Step2GuiHandler
 
 
 class CreateNdsumFile(object):
